Add files via upload
The project aims to develop a Python application that utilizes computer vision techniques to control the volume of a computer system through hand gestures. By leveraging the Mediapipe library for hand tracking and OpenCV for image processing, the application will enable users to adjust the volume by simply moving their hands up or down in front of a webcam.